HBSD: Start work on Cross-DSO CFI in base
This is the start of the work needed for Cross-DSO CFI in base. This is
very much a Work-In-Progress (WIP). With this commit, world does not
compile. However, all libraries are compiled with LTO, which is needed
for Cross-DSO CFI.

lld 4.0.0 is very strict about mixing -fpic and -fPIC. Some libraries
get compiled with -fpic instead of -fPIC. This causes linking issues in
base. It's likely that the PIC mode enforcement needs to be relaxed.

This also sets llvm-ar and llvm-nm as the default ar and nm (and
ranlib). elftoolchain doesn't understand LLVM IR bitcode, but llvm-ar
and llvm-nm do.

Also start work on porting libclang_rt.cfi, which is Linux-specific at
the moment. Note that the current code has not been tested since base
hasn't been successfully linked, yet.

Introduce a new toggle, MK_CROSS_DSO_CFI, which is enabled by default on
amd64 and disabled by default on all other architectures.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX